requirements.txt
python-dotenv==1.0.0
openai==1.3.5
tweepy==4.14.0
linkedin-api==2.0.3
newsapi-python==0.2.7
schedule==1.2.1
python-logging-loki==0.3.1
requests==2.31.0
prometheus-client==0.19.0
streamlit==1.28.2